Forms: move navigateToForm to a shared utility module
Decouple navigateToForm from convert-form-toolbar so it can be
imported without pulling in toolbar UI dependencies.
